eliminati tutti i valori short. Si usano solo i byte. In caso di conversione a int verificare eventuali errori sui valori massimi (255) Usare maschera "& 0xFF"

+++ org.openhab.binding.souliss/src/main/java/org/openhab/binding/souliss/handler/SoulissT12Handler.java
@@ -117,7 +117,7 @@ public class SoulissT12Handler extends SoulissGenericHandler {
                 case SoulissBindingConstants.SLEEP_CHANNEL:
                     if (command instanceof OnOffType) {
                         if (command.equals(OnOffType.ON)) {
-                            commandSEND((short) (SoulissBindingProtocolConstants.Souliss_T1n_Timed + xSleepTime));
+                            commandSEND((byte) (SoulissBindingProtocolConstants.Souliss_T1n_Timed + xSleepTime));
                             // set Off
                             updateState(channelUID, OnOffType.OFF);
                         }
